refact(payments): refactor payments module to DDD and isolate legacy debt types
